Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    int sum,a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    float sum,a,b,c;
+    scanf("%f %f %f",&a,&b,&c);
     sum=a+b+c/3;
-    printf("Average: %d",sum);
+    printf("Average: %.2f",sum);
     return 0;
 }
